Moved two functions to the plot utilities
And moved another one back to the test script.
